AI_projects/project1/pyproject.toml
[project]
name = "project1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.1.3",
    "langchain-openai>=1.1.2",
    "langgraph>=1.0.4",
    "python-dotenv>=1.2.1",
]
